Provide type information for generated unit parsers.
This adds a "TypeInfo" field to the parser table describing the
top-level unit type. It also provides host application with a 3rd
entry function for parsing that returns a new "ParsedUnit" instance
that generically wraps the parsed unit instance, and provides access
to its type information plus a raw pointer to its data. The host
application can use this as a starting point for iterating over all
the parsed fields (as our new spicy-dump will demonstrate in a
subsequent commit).

Technically we wouldn't need the type information inside the parser
table because the ParsedUnit coming out of the 3rd parse function
carries that already as well. However, might still be handy to have it
accessible directly without needing to trigger any parsing.